Ch2: Data Analysis, L3: Read File
Read File

Let's start ackshually building "bookbot"!
Here are some things you'll need to know how to do for this step:

with Block
https://docs.python.org/3/reference/compound_stmts.html#with

A with block can be used to open a file:
https://docs.python.org/3/tutorial/inputoutput.html#
  reading-and-writing-files

```python
with open(path_to_file) as f:
    # do something with f (the file) here
```

The with block will automatically close the file
when the block is exited, cleaning up resources.

---------------------------

.read() Method
https://python-reference.readthedocs.io/en/latest/docs/file/read.html

You can use the .read() method to read the contents
of a file into a string:

file_contents = f.read()

Assignment

- Delete your "greetings boots" message.
- Write a new function called get_book_text. It takes a filepath
  as input and returns the contents of the file as a string.
- Write a main function that uses get_book_text with the relative path
  to your frankenstein.txt file to print the entire contents of the book
  to the console.
- Call main() at the bottom of your file to execute your program.
- Test your program!

python3 main.py

Run and submit the CLI tests.

Tips

- The path to a file is a string, so it needs to be in quotes.
- You don't need to write all your functions and variable names the
  same as mine, I'm just trying to give you some examples of how you
  can potentially structure your code. There are many correct ways
  to write the same program.