Ch2:Data Analysis, L4: Count Words
Count Words

Time for our first bit of data analysis!
Let's count the words in the book.

Assignment

- Write a new function that accepts the text from the book as a string,
  and returns the number of words in the string. The .split() method
  will be helpful here.
- Instead of printing the books contents, print this message
  to the console:

  {num_words} words found in the document

Replace {num_words} with the actual number of words in the book.

Run and submit the CLI tests.

--------------
90% My solution

The main difference is that I did originally return the f
string from word_count. I changed this when I noticed Boot.dev's
solution didn't do this and was like "duh, of course we should
just return the word count from `def word count` of course"