Ch2: Data Analysis, L5: Refactor
Refactor

Refactoring is a fancy word for
"rewriting your god-awful code so it's not so terrible."

Refactoring doesn't change the behavior of your code,
just the way it's written.

Let's refactor our project to use multiple Python files.

Not only will it be useful for you to learn how to split code
across files, but in this case, it will help us organize our code.
We'll use this structure:

- main.py - The entry point to our program and any code that doesn't
  fit elsewhere
- stats.py - A file that contains functions for analyzing the text

-----------------------------------------------
Assignment

- Move your function that counts the words
  into a new stats.py file.
- Import the function in your main.py file,
  it will look something like this:

```
from stats import get_num_words
```

Where stats is the name of the file (without the .py extension) and
get_num_words is the name of the function you want to import.

Run and submit the CLI tests.

----------------------------------------------
Boot.dev's solution is close enough to mine (just minor differences)