Ch3: Report, L1: Print a Report (Boot.dev's fix)
- order imports on separate lines
- iterate over a dictionary in a more pythonic way
  (than what I was doing)
- extract report printing from `main` into a `print_report` method

NOTES:
1. This isn't exactly Boot.dev's fix. Instead, it updates
my fix to be more similar to Boot.dev's. With the changes
here it more or less is Boot.dev's fix with the exception
of method ordering and a few variable/function names.

2. I don't really like the continue being used here, however
since I could see being able to use a continue in python code
being useful I updated the code to use it so that I have an
example of it incase I need to grok it in the future.