Ch3 Report, L2: Arguments (my fix)
Arguments

Bookbot is done! Almost...

You've probably noticed that most CLI tools you're used to
(ls, cd, git, etc.) don't work with a single "hard coded" input.
Right now bookbot can only analyze books/frankenstein.txt.
That's not very useful, is it?

System Arguments

The built in sys module provides access to command line arguments.
In particular sys.argv is a list of strings representing the arguments
passed to the script. The first argument is the script name, the rest
are the arguments.

To import entire built-in modules like sys, use
the standard import syntax: import sys.

For example, python3 main.py will result in a sys.argv list
that looks like this:

['main.py']

And python3 main.py books/frankenstein.txt will result in
a sys.argv list that looks like this:

['main.py', 'books/frankenstein.txt']

Assignment

1. Remove any hard coded book paths from your code.
2. At the beginning of main.py, import the built-in sys module.
3. If sys.argv doesn't have two entries:
  - Print a message explaining how to use the program:
    Usage: python3 main.py <path_to_book>

  - Exit the program with a status code of 1 using sys.exit(1)

4. Use the second argument in sys.argv as the path to the book file.
5. Test your program with python3 main.py books/frankenstein.txt and
   make sure it works as expected.

6. Download some more books from Project Gutenberg and test your program
   with them, adding them to the books directory and passing them in as
   CLI arguments when you run your program.
   Add at least these two books:

   - books/mobydick.txt
   - books/prideandprejudice.txt

Use the books that are linked above, or the CLI tests will fail.

Run and submit the CLI tests.

NOTE: Boot.dev's solution more or less matches. The only real
difference is that they didn't check if the number of CLI arguments were
exactly 2 they checked `if len(sys.argv) < 2:`. Which I don't think
is as good as mine because if the user specifies 3 or more arguments
they might be expecting something different to happen and it doesn't.
So, I think 2 is better in communicating what the program expects.